+++ src/integratere.cpp
@@ -21,7 +21,9 @@ using namespace Rcpp;
 //'   0 indicates inverse logit (e.g., for logistic regression).
 //'   1 indicates exponential (e.g., for poisson or negative binomial regression or if outcome was natural log transformed).
 //'   2 indicates square (e.g., if outcome was square root transformed).
-//'   Any other integer may be used for no transformation.
+//'   3 indicates inverse (e.g., if outcome was inverse transformed such as Gamma regression)
+//'   Any other integer results in no transformation. -9 is recommended as the option for no
+//'   transformation as any future transformations supported will be other, positive integers.
 //' @return A numeric matrix with the Monte Carlo integral calculated.
 //' @export
 //' @examples
@@ -72,6 +74,8 @@ arma::mat integratere(List d, List sd, List L, int k, const arma::mat& yhat, int
       Zall = exp(Zall);
     } else if (backtrans == 2) {
       Zall = pow(Zall, 2);
+    } else if (backtrans == 3) {
+      Zall = 1 / Zall;
     }
     arma::colvec zm = arma::mean(Zall, 1);
     yhat2.row(i) = zm.t();