+++ src/lmcpp.h
@@ -1,0 +1,9 @@
+#ifndef __lmcpp__
+#define __lmcpp__
+
+#include <RcppArmadillo.h>
+arma::mat lmcpp(const arma::mat& X, const arma::mat& y);
+
+#endif // __lmcpp__
+
+