+++ tests/testthat/test-dot-links.R
@@ -5,6 +5,7 @@ test_that(".links returns correct values with identity link and fixedonly", {
   expect_type(x, "list")
   expect_equal(x$scale, "response")
   expect_equal(x$ilink, "identity")
+  expect_equal(x$useifun(0.5), 0.5)
   expect_equal(x$ilinknum, -9)
 })
 
@@ -14,8 +15,10 @@ test_that(".links returns correct values with logit link and fixedonly", {
 
   expect_type(x, "list")
   expect_equal(x$scale, "response")
-  expect_equal(x$ilink, "identity")
-  expect_equal(x$ilinknum, -9)
+  expect_equal(x$ilink, "invlogit")
+  expect_equal(x$useifun(0.5), 0.5)
+  expect_equal(x$ilinknum, 0L)
+  expect_equal(x$useilinknum, -9L)
 })
 
 test_that(".links returns correct values with identity link and integrateoutRE", {
@@ -25,6 +28,7 @@ test_that(".links returns correct values with identity link and integrateoutRE",
   expect_type(x, "list")
   expect_equal(x$scale, "linear")
   expect_equal(x$ilink, "identity")
+  expect_equal(x$useifun(0.5), 0.5)
   expect_equal(x$ilinknum, -9)
 })
 
@@ -35,7 +39,10 @@ test_that(".links returns correct values with logit link and integrateoutRE", {
   expect_type(x, "list")
   expect_equal(x$scale, "linear")
   expect_equal(x$ilink, "invlogit")
-  expect_equal(x$ilinknum, 0)
+  expect_equal(x$ifun(0.5), plogis(0.5))
+  expect_equal(x$useifun(0.5), 0.5)
+  expect_equal(x$ilinknum, 0L)
+  expect_equal(x$useilinknum, 0L)
 })
 
 test_that(".links returns correct values with log link and integrateoutRE", {
@@ -45,7 +52,10 @@ test_that(".links returns correct values with log link and integrateoutRE", {
   expect_type(x, "list")
   expect_equal(x$scale, "linear")
   expect_equal(x$ilink, "exp")
-  expect_equal(x$ilinknum, 1)
+  expect_equal(x$ifun(0.5), exp(0.5))
+  expect_equal(x$useifun(0.5), 0.5)
+  expect_equal(x$ilinknum, 1L)
+  expect_equal(x$useilinknum, 1L)
 })
 
 test_that(".links returns correct values with sqrt link and integrateoutRE", {
@@ -55,5 +65,8 @@ test_that(".links returns correct values with sqrt link and integrateoutRE", {
   expect_type(x, "list")
   expect_equal(x$scale, "linear")
   expect_equal(x$ilink, "square")
+  expect_equal(x$ifun(0.5), (0.5)^2)
+  expect_equal(x$useifun(0.5), 0.5)
   expect_equal(x$ilinknum, 2)
+  expect_equal(x$useilinknum, 2)
 })