+++ tests/testthat/test-predict-mixedlogit.R
@@ -81,7 +81,7 @@ res.fixedonly <- withr::with_seed(
         test0$Posterior, test1$Posterior, ame$Posterior))
   })
 
-test_that(".predict works to integrate out random effects in multilevel logistic models", {
+test_that("prediction works to integrate out random effects in multilevel logistic models", {
   expect_type(res.integrate, "list")
   expect_equal(
     c(ndraws(mlogit), 3L),
@@ -99,7 +99,7 @@ test_that(".predict works to integrate out random effects in multilevel logistic
                     (res.samp$M[2] - res.samp$M[1])) < .01)
 })
 
-test_that(".predict works with fixed effects only in multilevel logistic models", {
+test_that("prediction works with fixed effects only in multilevel logistic models", {
   expect_type(res.fixedonly, "list")
   expect_equal(
     c(ndraws(mlogit), 3L),